implement valid LeNet5 model add LeNet5 handling in main add calculation of precision and recall for each class in model

+++ src/oct_data_generator.py
@@ -39,7 +39,8 @@ class OCTDataGenerator(keras.utils.Sequence):
             np.random.shuffle(self.indexes)
 
     def _data_generation(self, item_paths):
-        x = np.empty((self.batch_size, *self.dim, self.n_channels), dtype=np.float16)
+        # TODO: change to lower precision (ex. float16) if needed
+        x = np.empty((self.batch_size, *self.dim, self.n_channels), dtype=np.float64)
         y = np.empty(self.batch_size, dtype=int)
         for counter, filepath in enumerate(item_paths):
             img_arr = cv2.imread(filepath, cv2.IMREAD_GRAYSCALE) / 255.0
